+++ crates/ruma-client-api/CHANGELOG.md
@@ -4,6 +4,8 @@ Improvements:
 
 - Add `uiaa::LoginTermsParams`, that allows to construct or extract the
   parameters of the `m.login.terms` authentication type.
+- Add `UiaaInfo::params()` as a helper to extract UIAA authentication type
+  parameters.
 
 # 0.22.0