+++ crates/ruma-client-api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # [unreleased]
 
+Improvements:
+
+- Add `uiaa::LoginTermsParams`, that allows to construct or extract the
+  parameters of the `m.login.terms` authentication type.
+- Add `UiaaInfo::params()` as a helper to extract UIAA authentication type
+  parameters.
+
 # 0.22.0
 
 Breaking changes: